pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speechlab-diarization"
version = "0.1.0"
description = "Speaker diarization with voice-type classification using pyannote and VTC 2.0"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "CS26-05 SpeechLab Team"}
]
keywords = [
    "speech",
    "diarization",
    "speaker-diarization",
    "voice-type-classification",
    "pyannote",
    "vtc",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "pyannote.audio>=3.1,<4.0",
    "torch>=2.0",
    "torchaudio>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
speechlab-diarize = "speechlab_diarization.main:main"

[project.urls]
Repository = "https://github.com/CS26-05/SpeechLab"

[tool.setuptools.packages.find]
where = ["."]
include = ["speechlab_diarization*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

